Extracted test_suite_fxiture_to_case_selection
